persistence: add ranking support
Implement two different ways of ranking.

One pushes the work completely to MongoDB using the aggregation framework.
This requires actually a fair amount of in-memory work on the server
side and would not be a good idea for large collections.

Another simply has Mongo sort the collection and then scans a cursor,
stopping when we've found the appropriate rank. This should be
reasonably efficient as Mongo (purportedly) does batching under the
hood, and we can also eventually put indices on these documents
so that sorting is amortized during insertions.

See https://stackoverflow.com/questions/44587829/sorting-and-ranking-documents-based-on-key-values