app: working wire-up of new and old database libs
Begin using the persistence.js code. Start with using it to connect and
load up repositories.  To shim in the old code, we inject the database
object from persistence.js into leaderboard_fetch.

Start by using the UserRepository when creating user on message read, but
the old code for everything else.