FEATURE: automatic VC creation.
Creates a new voice channel called "Extra Practice Room" if all registered PRs are occupied by either muted
or unmuted users. Muted or unmuted doesn't matter because we want to discourage people from using rooms that
are occupied even if all the participants are currently muted - it could be the case that someone was practising
there but was just muted temporarily.

If the practice rooms *aren't* full, the bot checks to see if it should remove one of the temp rooms. It will
remove a temp room if there are at least two empty rooms, and one of those empty rooms is a temp room. (If none
of the empty rooms is a temp room, then the only rooms we can remove are the main rooms, which we don't want to
remove. If there is only one empty room, we need that room to be available.)