BUG: 36 ignore ghost sessions in VC auto-creation and live leaderboard code (#50)
Due to a Discord bug, users who leave the server while in a VC never emit the
voiceStateUpdate event to tell us they're leaving. As a result, the user stays
in the channel, accumulates time if unmuted, and counts as occupying the room
for the purposes of VC autocreation. This change ignores GuildMember objects
that have the `deleted` flag set to `true`.