Implement p!lock and p!unlock
Implements user commands to take and release exclusive control of
practice rooms. A few API complications informed the design:

- `p!lock` checks if the user is currently in a channel and locks
the channel if not already locked (assuming the channel is a
registered practice room). Locking a room involves changing
permissions on the room such that @everyone is denied the SPEAK
permission, and an override is added for the occupant to allow the
SPEAK permission. Additionally, everybody else in the room becomes
server muted (this is because denying the SPEAK permission does
not mute users who are currently in the room. The reason why we
don't solely rely on server mutes is because there may be a delay
between when an unmuted user enters a locked room and when a server
mute kicks in.)

- `p!unlock` checks if the user is in a channel locked by them,
then unlocks the room by undoing the permissions changes and server
unmuting all users (if they were already server unmuted, this still
has the effect of unmuting users who are muted by virtue of having
been denied the SPEAK permission). Bot Managers can unlock rooms
they don't have locked themselves (or rooms they're not in).

- the time calculation is modified so that we don't start the timer
if the room is locked by someone else - this is because being muted
by virtue of permission does not actually reflect in the `mute` and
`selfMute` properties of `GuildMember`. Without this change, a user
who joins a locked room and is yet server unmuted would be live.

- the occupant leaving a locked room automatically unlocks the room.

- anybody else leaving a locked room should be unmuted. However...
if they aren't in a channel anymore, we can't actually modify their
muted/unmuted status. Hence, we unmute them if they ever join an
unlocked channel. One concern, of course, is that they may be
server muted for other reasons - i.e. manually, by a moderator. As
a result, the bot will never unmute users with the Temp Muted role.