BUG: 61 Server-muted users leaving locked room retains PR chat privileges
The "should we unmute the user as they are now in an unlocked room"
check didn't check to make sure they were actually in a room, so in
the case they weren't (i.e. left the room entirely), the event handler
bailed completely and didn't continue to the PR chat permissions update.