Remove dev_mode and beta_ settings
Eliminate the concept of dev_mode so that there can be no opportunity
for us to introduce code that is contingent on the concept of dev vs
production. Instead, move configuration for key parameters into
environment variables.

This means that developers can set a local variable to their test bot
tokens, etc without having to modify the settings.json file in the repo.