session: create concept of practicing
Consolidate various places where we check to see that a user
has a session long enough to be considered "practicing". Extract
this limit (15 minutes) to a setting (for easier testing) and
then use that concept throughout.

When a user is practicing (as opposed to "just" starting a session, for
which we still track time), we mark them as having practiced on their
profile and give them credit for practicing in a given room.