Nerf gifted monkey rooms
Monkey rooms latch the first user to enter the room. If that user
leaves, the monkey disappears. If the user sticks around, another user
cannot claim the room nevertheless. The first user has fifteen minutes
to claim the room before the monkey disappears (if the user is already
in the process of claiming the room, then this does not apply). In all
cases, if the monkey is not successfully claimed, it goes back onto the
token queue immediately.

The net result is that monkeys must be claimed immediately (or close to
immediately) by the FIRST user to join a room.

New badge added to compensate for the increase in difficulty.