Make monkeys possible to obtain outside of randomly, but extremely difficult (#193)
Spawns monkey tokens in empty rooms instead of occupied rooms. This makes it possible for users to actively hunt them instead of being gifted them by the RNG, but to compensate, the strategy of having friends hold monkeys is nerfed.

More specifically, monkeys want one of two things: they either want to be _claimed_, or to be _free_. The first user to enter an empty monkey room latches the room. Another user cannot claim the token, even after the user leaves the room. (This status resets if the room becomes empty again, but then again, room autodeletion may just wipe the room entirely.)

To prevent a user from holding a monkey room and preventing other monkeys from spawning, a monkey escapes within fifteen minutes if the token has not been claimed and someone is not in the process of trying to claim the token.

To prevent a user from pursuing a strategy of intentionally despawning monkey rooms (either by latching a room and making the monkey unclaimable, or despawning via room autodeletion, or letting the monkey run away), _any monkey token that goes unclaimed is immediately returned to the top of the token queue_. It will automatically be the next token spawned on room autocreation, regardless of whether the deleted token stack has timed out or not. If no room autocreation happens, then the monkey is spawned normally, but with 100% probability at the next available opportunity.

The result is that 1) monkeys must be claimed immediately (or close to immediately) by the first user to join a room and 2) if that user is unable to practice, then the monkey will free itself as soon as possible so that someone else can claim it.

New badge added to compensate for the increase in difficulty.